LU-12678 lnet: remove locking protection ln_testprotocompat
lnet_net_lock(LNET_LOCK_EX) is a heavy-weight lock that is not
necessary here.  The bits in this field are only set rarely - via an
ioctl - and the pattern for reading and clearing them exactly
matches test_and_clear_bit().  So change the field to "unsigned
long" (so test_and_clear_bit() can be used), and use
test_and_clear_bit(), discarding all other locking.

Test-Parameters: trivial testlist=sanity-lnet
Reviewed-on: https://review.whamcloud.com/36856
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>

+++ lnet/include/lnet/lib-types.h
@@ -1113,7 +1113,7 @@ struct lnet {
 	const struct lnet_lnd		*ln_lnds[NUM_LNDS];
 
 	/* test protocol compatibility flags */
-	int				ln_testprotocompat;
+	unsigned long			ln_testprotocompat;
 
 	/* 0 - load the NIs from the mod params
 	 * 1 - do not load the NIs from the mod params
